added stylus, Hello -> Home
