+++ package.json
@@ -17,6 +17,8 @@
   },
   "dependencies": {
     "buefy": "^0.5.2",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.1",
     "vue": "^2.4.2",
     "vue-resource": "^1.3.4",
     "vue-router": "^2.7.0",
