+++ src/App.vue
@@ -1,6 +1,5 @@
 <template>
-  <div id="app">
-    <img src="./assets/logo.png">
+  <div id="app" class="container">
     <router-view></router-view>
   </div>
 </template>
@@ -11,15 +10,12 @@
   }
 </script>
 
-<style>
-  @import "~buefy/lib/buefy.css";
-  @import "~bulma/css/bulma.css";
-  #app {
-    font-family: 'Avenir', Helvetica, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #2c3e50;
-    margin-top: 60px;
-  }
+<style lang="stylus">
+  @import "~buefy/lib/buefy.css"
+  @import "~bulma/css/bulma.css"
+  #app
+    font-family: Helvetica, Arial, sans-serif
+    -webkit-font-smoothing: antialiased
+    -moz-osx-font-smoothing: grayscale
+    margin-top: 60px
 </style>
