--- src/components/Hello.vue
@@ -1,105 +1,0 @@
-<template>
-  <div class="hello">
-    <h1>{{ msg }}</h1>
-    <h2>Essential Links</h2>
-    <ul>
-      <li><a href="https://vuejs.org" target="_blank">Core Docs</a></li>
-      <li><a href="https://forum.vuejs.org" target="_blank">Forum</a></li>
-      <li><a href="https://chat.vuejs.org" target="_blank">Community Chat</a></li>
-      <li><a href="https://twitter.com/vuejs" target="_blank">Twitter</a></li>
-      <br>
-      <li><a href="http://vuejs-templates.github.io/webpack/" target="_blank">Docs for This Template</a></li>
-    </ul>
-    <h2>Ecosystem</h2>
-    <ul>
-      <li><a href="http://router.vuejs.org/" target="_blank">vue-router</a></li>
-      <li><a href="http://vuex.vuejs.org/" target="_blank">vuex</a></li>
-      <li><a href="http://vue-loader.vuejs.org/" target="_blank">vue-loader</a></li>
-      <li><a href="https://github.com/vuejs/awesome-vue" target="_blank">awesome-vue</a></li>
-    </ul>
-    <h2>Some Resources</h2>
-    <b-table :data="someResources">
-
-      <template scope="props">
-          <b-table-column label="RID" width="60" numeric>
-              {{ props.row.rid }}
-          </b-table-column>
-
-          <b-table-column label="Name">
-              {{ props.row.rname }}
-          </b-table-column>
-      </template>
-
-      <template slot="empty">
-          <section class="section">
-              <div class="content has-text-grey has-text-centered">
-                  <p>
-                      <b-icon
-                          icon="sentiment_very_dissatisfied"
-                          size="is-large">
-                      </b-icon>
-                  </p>
-                  <p>Nothing here.</p>
-              </div>
-          </section>
-      </template>
-    </b-table>
-  </div>
-</template>
-
-<script>
-  import { mapGetters, mapActions } from 'vuex'
-  export default {
-    name: 'hello',
-    data () {
-      return {
-        msg: 'Welcome to Your Vue.js App'
-      }
-    },
-    computed: {
-      ...mapGetters([
-        'someResources'
-      ])
-    },
-    methods: {
-      ...mapActions([
-        'fetchResource'
-      ]),
-      async fetch () {
-        this.busy = true
-        await this.fetchResource({
-          url: '/api/v1/some-resources',
-          prop: 'someResources'
-        })
-        this.busy = false
-      }
-    },
-    async created () {
-      this.fetch()
-    },
-    mounted () {
-      console.log('* Mounted Hello Component!')
-    }
-  }
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-  h1, h2 {
-    font-weight: normal;
-  }
-
-  ul {
-    list-style-type: none;
-    padding: 0;
-  }
-
-  li {
-    display: inline-block;
-    margin: 0 10px;
-  }
-
-  a {
-    color: #42b983;
-  }
-</style>
